icon change

+++ lib/app/models/home/controllers/image_picker_controller.dart
@@ -1,15 +1,14 @@
 import 'dart:io';
 
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image_cropper/image_cropper.dart';
+import 'package:image_gallery_saver/image_gallery_saver.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:pdf/widgets.dart' as pw;
 import 'package:hive/hive.dart';
 import 'package:share_plus/share_plus.dart';
-
 
 import '../../../core/values/s_datetime_util.dart';
 
@@ -19,7 +18,8 @@ class ImagePickerController extends GetxController {
   final pdfPath = ''.obs; //path of generated pdf
   final loading = false.obs; // Loading state observable
   var enlargedImagePath = ''.obs;
- final imageDates =<String,String>{}.obs;//Map to store image path and their dates
+  final imageDates =
+      <String, String>{}.obs; //Map to store image path and their dates
 
   final ImagePicker _picker = ImagePicker();
 
@@ -48,7 +48,7 @@ class ImagePickerController extends GetxController {
       final XFile? image = await _picker.pickImage(source: ImageSource.camera);
       if (image != null) {
         imagePaths.add(image.path);
-           imageDates[image.path] = SDateTimeUtil.dateTimeToString(
+        imageDates[image.path] = SDateTimeUtil.dateTimeToString(
           DateTime.now(),
           SDateTimeUtil.formatPattern6,
         );
@@ -60,11 +60,9 @@ class ImagePickerController extends GetxController {
     }
   }
 
-
   void toggleSelection(int index) {
     if (selectedIndexes.contains(index)) {
       selectedIndexes.remove(index);
-
     } else {
       selectedIndexes.add(index);
     }
@@ -81,37 +79,50 @@ class ImagePickerController extends GetxController {
         );
       }));
     }
-    
+
     try {
       final dir = await getApplicationDocumentsDirectory();
       final file = File('${dir.path}/selected_images.pdf');
       await file.writeAsBytes(await pdf.save());
       pdfPath.value = file.path;
+      await savePdfPath(file.path);// Save PDF path to Hive
       Get.snackbar('Success', 'PDF created ');
-    } catch(e){
+    } catch (e) {
       print(e);
     }
   }
 
-  Future<void> saveSelectedImagesToHive() async {
-    if(selectedIndexes.isNotEmpty){
-      final box = await Hive.openBox<String>('selectedImagesBox');
-    for (var index in selectedIndexes) {
-      await box.add(imagePaths[index]);
+  Future<void> savePdfPath(String path) async{
+    final box = await Hive.openBox<String>('pdfPathBox');
+    if(await File(path).exists()){
+      if(!box.values.contains(path)){
+        await box.add(path);
+      }
     }
-    Get.snackbar('Success', 'Selected images saved ');
-    selectedIndexes.clear();
+  }
+
+  Future<void> saveSelectedImagesToGallery() async {
+    if (selectedIndexes.isNotEmpty) {
+      for (var index in selectedIndexes) {
+        String imagePath = imagePaths[index];
+        final result = await ImageGallerySaver.saveFile(imagePath);
+        if(result['isSuccess']){
+          print('Image Saved to gallery: $imagePath');
+        }else{
+          print('Failed to save image to gallery: imagePath');
+        }
+      }
+      Get.snackbar('Success', 'Selected images saved ');
+      selectedIndexes.clear();
     }
-   
   }
 
   Future<void> saveImagePath(String path) async {
     final box = await Hive.openBox<String>('imagePathsBox');
     if (await File(path).exists()) {
-      if(!box.values.contains(path)){
-       await box.add(path);
+      if (!box.values.contains(path)) {
+        await box.add(path);
       }
-      
     }
   }
 
@@ -122,33 +133,32 @@ class ImagePickerController extends GetxController {
     imagePaths.addAll(validPaths);
   }
 
- Future<void> deleteSelectdItems() async{
-    final imagePathBox = await Hive.openBox<String>('imagePathBox');
-    List<String> pathsToDelete = selectedIndexes.map((index) => imagePaths[index]).toList();
+  Future<void> deleteSelectdItems() async {
+    if (selectedIndexes.isNotEmpty) {
+      final imagePathBox = await Hive.openBox<String>('imagePathBox');
+      List<String> pathsToDelete =
+          selectedIndexes.map((index) => imagePaths[index]).toList();
 
       for (String path in pathsToDelete) {
-      final file = File(path);
-      if (file.existsSync()) {
-        file.deleteSync();
-      }
-      //Find and delete the path from HIve
-      final keyToDelete = 
-         imagePathBox.keys.cast().firstWhere((key) => imagePathBox.get(key!) == path,orElse: () => null,);
-         if(keyToDelete!=null){
-              await imagePathBox.delete(keyToDelete);
-            
-         }
-     
+        final file = File(path);
+        if (file.existsSync()) {
+          file.deleteSync();
+        }
+        //Find and delete the path from HIve
+        final keyToDelete = imagePathBox.keys.cast().firstWhere(
+              (key) => imagePathBox.get(key!) == path,
+              orElse: () => null,
+            );
+        if (keyToDelete != null) {
+          await imagePathBox.delete(keyToDelete);
+        }
       }
 
-       imagePaths.removeWhere(
-        (path) => pathsToDelete.contains(path));
-    selectedIndexes.clear();
-    
-      Get.snackbar('Success','Deleted successfully');
+      imagePaths.removeWhere((path) => pathsToDelete.contains(path));
+      selectedIndexes.clear();
 
-    
-   
+      Get.snackbar('Success', 'Deleted successfully');
+    }
   }
 
   Future<void> flushHiveMemory() async {
@@ -159,28 +169,29 @@ class ImagePickerController extends GetxController {
     imagePaths.clear();
     selectedIndexes.clear();
     pdfPath.value = '';
-    Get.snackbar('Success', 'Hive memory flushed');
+    Get.snackbar('Success', 'Deleted saved memory');
   }
 
-   void shareSelectedImages() {
+  void shareSelectedImages() {
     if (selectedIndexes.isNotEmpty) {
-     List<String> selectedImagePaths = selectedIndexes.map((index)=>imagePaths[index]).toList();
+      List<String> selectedImagePaths =
+          selectedIndexes.map((index) => imagePaths[index]).toList();
       selectedImagePaths.map((path) => XFile(path)).toList();
-     Share.shareXFiles([XFile(pdfPath.value)], text: 'Here is the PDF of images.');
+      Share.shareXFiles([XFile(pdfPath.value)],
+          text: 'Here is the PDF of images.');
     } else {
       Get.snackbar('Error', 'No PDF available to share');
     }
   }
 
-  void enlargeImage(String path){
-   enlargedImagePath.value = path;
+  void enlargeImage(String path) {
+    enlargedImagePath.value = path;
   }
 
   Future<void> cropImage() async {
     if (enlargedImagePath.value.isNotEmpty) {
-      CroppedFile? croppedFile =await ImageCropper().cropImage(
-        sourcePath: enlargedImagePath.value,
-       uiSettings: [
+      CroppedFile? croppedFile = await ImageCropper()
+          .cropImage(sourcePath: enlargedImagePath.value, uiSettings: [
         AndroidUiSettings(
           toolbarTitle: 'crop Image',
           toolbarColor: Get.theme.colorScheme.primary,
@@ -188,9 +199,7 @@ class ImagePickerController extends GetxController {
           initAspectRatio: CropAspectRatioPreset.original,
           lockAspectRatio: false,
         ),
-       ]
-  
-      );
+      ]);
 
       if (croppedFile != null) {
         // Update the image path in the imagePaths list
@@ -204,7 +213,6 @@ class ImagePickerController extends GetxController {
       }
     }
   }
-
 
   @override
   void onInit() {
